Document how to use Lua 5.4 and power patches
This commit adds documentation on how to use Lua 5.4 and documents the
power patches provided by this new release. Most of the documentation
for this comes from the readme (https://github.com/gottfriedleibniz/lua/blob/luaglm/README.md),
'as is' and the lua-users Wiki: http://lua-users.org/wiki/LuaPowerPatches#LUA_54_patches
I tried to not modify any of the documentation, only adding certain
markdown styling and some extra explanation to specific terms which may not be
known to new users.