Create documentation contribution FAQ
I created this section to try and address a request where a user
correctly claimed that there's missing documentation. The main goal
behind this FAQ is emphasizing on the importance of the Github issues
section for both natives and docs.

It's difficult to know what's higher in priority to document when
there's users with different needs, emphasizing on creating issues for
documentation and natives will hopefully help with this, especially if
multiple users come up with a common goal.