Add crosshair research by Fengorian
I made some changes in the layout so it follows how 'Console commands' are displayed to the end user.

Also added documentation for `cl_crosshair_outlinethickness`, which was
missing, as well as certain nitpicks at `cl_crosshaircolor`, which should be set
to `5` for custom crosshair colors to work.

Explain available color presets

Explain what the available color presets are, as in those applied via
`cl_crosshaircolor`.

Get rid of unneeded new line.